fix(stella-cli): split the loop's reporting verbs out of self_driving_cmd.rs
`crates/stella-cli/src/self_driving_cmd.rs` reached 1501 lines and is not
grandfathered, so the file-size ratchet fails on the merged tree. AGENTS.md
is explicit about the remedy for a first-time crossing: split it into a
sibling submodule, never give it a baseline entry.

`watch` and `metrics` move to `self_driving_cmd/report.rs`. They were
already one section behind a shared `// watch / metrics` banner, so the seam
is the one the file had drawn for itself rather than a cut chosen to hit a
line count. Both are read-only reporting verbs: they answer a question about
the loop without moving it, which is what makes them a module rather than an
arbitrary 123 lines. `watch`'s one write is the aperture reset a wake
implies — the observation becoming the decision, not a side effect of
reporting — and that is said out loud in the new module's header.

No visibility was widened. A child module can already see its parent's
private items, so `say`, `gh_available`, `gh_plain` and `demand` stay private
and are reached through `super::`.

The witness is the guard itself, run the way the canary runs it:

  - on `main`, `./scripts/check-file-size.sh --absolute` names
    self_driving_cmd.rs at 1501 lines and fails;
  - here it exits OK on the strict whole-tree check.

The file lands at 1378 lines, so the headroom is real rather than one line
under the wire. Behaviour is unchanged: the two functions move verbatim and
their call sites in the dispatch `match` are re-pointed at `report::`.

Closes #4044
